requirements.txt
qdrant-client>=1.9.0
streamlit>=1.35.0
python-dotenv>=1.0.1

transformers>=4.30.0
torchvision>=0.17.0
timm>=0.9.0
sentence-transformers>=2.2.2
scikit-learn>=1.2
pandas>=2.0
accelerate>=0.20.3
